Use the NixOS module system for out-of-tree package definitions
Nixpkgs is a monorepo, and its design pattern makes it hard to use in polyrepo contexts.

This change allows out-of-tree packages to be defined using the NixOS module system, a system designed for modularity.

Multiple regular, Python and ROS packages can be added through module options.

This system also leaves room for packages to make workspace customisations, if that is needed in the future.