reolink: use shell script instead of aliases
not sure how to get this set up correctly with nix - for now
i have a messy solution where an alias runs the script directly
from the git repository

make sure you are in a `nix-shell -p ffmpeg`. Then you can run
`reolink` or `reolink low` or `reolink 180` or `reolink low 180` or
`reolink 180 low` etc etc. Argument order doesn't matter. 180 is for if
the camera is upside down.