reolink-ctl: try to add speed control for auto mode
its not very good

we control speed by sending a new move command when the camera is moving
this makes the camera stop then start moving. If you do this frequently,
it is really slow, if you do it infrequently it is normal speed but it
moves in bursts.

Use keys 1234 to modify the two parameters for this.

if you want the previous mode, use speed=6, pan duration=6 i thkin