nova_arm_controller: basic vel and pos control
used the JointSaturationLimiter to limit position and velocity

need to actually collect current pos/vel (and calculate acc/jerk) to get
this working properly, currently I just tell it that we are always at
position zero.

The limits are configured in nova_arm_controller.yaml. I have made it so
J1 has velocity limit of 0, and that made J1 stop moving :)