cmd_hardware: move resolver discontinuity to +-180deg
In the case of j6 specifically, this means that if you start the
hardware interface while it is within 45 degrees of the zero position
its good. Without this, if you were for example at -1 degree, it would
think j6 is at 44 degrees (because of the 4 as the gear ratio).

more context:

The resolver returns a 16 bit unsigned value from 0 to 0x3fff (or you
can consider this a signed/unsigned 14 bit value with two added 0's).

Currently the jump between 0 and 0x3fff is at 0degrees, the zero
position. We change how the code interprets this so that it is effectively
a 14 bit signed integer and the discontinuity moves to 0x1fff-0x2000
which is at 180 degrees.

This is analogous to what the old arm code does. The old arm code will
assume that upon startup or zeroing that the joint is near the actual
zero, so if it gets something from 0 to 180 degrees it will assume it
is in sector 0, if it gets something in 180-360 degrees it will assume it
is in sector (gear_ratio-1). See reset_sector_count in
resolver_publisher.py for the old implementation.

There is also some code in the old arm resolver stuff that is for moving
discontinuities away from each joint's main operating area, but it
seemed like it just moves the discontinuity for j6 to 0 degrees so idk.
This moving of discontinuities happens after the sector count and gear
ratio stuff is already processed.

$ ros2 service call arm/arm_config_info arm_interfaces/srv/ArmConfigInfo
joint_names=['arm_j1', 'arm_j2', 'arm_j3', 'arm_j4', 'arm_j5', 'arm_j6']
joint_limits_lower=[-2*pi, -1.75, -2.75, -1.95, -1.95, -2*pi]
joint_limits_upper=[ 2*pi,  1.55,  2.40,  1.57,  1.75,  2*pi]

discontinuity_angle = (limit_upper+limit_lower) % (2*pi)
move_discontinuity(angle, discontinuity_angle):
  return angle - 2*pi * (angle >= discontinutiy_angle)

for j6 the discontinuity_angle is 0 so move_discontinutiyy always
subtracts 2pi from the position readings in this case. I don't think
that's important for us. The same is true for j1, but j2,3,4,5 might be
impacted by this if you want to go for example +200 to -160 instead of
from +180 to -180 with one of these joints. Looking at the arm in rviz I
don't think we'll want any joints moving more than +-180 degrees from zero
position anyway.