feat: can.sh: canable support
make sure the can script works with canable. It does not support the
berr-reporting option, so don't error if that can't be used.

I have removed the manual loading of kernel modules as these will be
automatically loaded when the kernel enumerates a device with a matching
modailas to the modules. If this is not the case then the module is not
installed properly (in which case modprobe wouldn't be able to find them
in the first place so it shouldn't work anyway).