feat: can.sh: can mv to rename a can bus
Use can mv [src] [dst] to rename a can bus. Eg `can mv can0 can1`. There
will be an alias kept indicating what the name of each moved bus was
before it was moved the first time, eg "can0_orig". You can see this in
the output of `ip link` for example.

Use this if you want to run arm on can0, or swap the canable to be named
can1.