fix: nova-git-metadata: don't make builds fail without .git
When hydra tries to build this it fails because .git is missing. This
means hydra doesn't build this and the final workspace package.

Check if .git is missing and if so then just fail silently to ensure it
doesn't break the build.

see also: https://github.com/MonashNovaRover/nova/pull/735#discussion_r2442434487